+++ src/components/textarea/textarea.js
@@ -58,12 +58,22 @@ class Textarea extends React.Component {
      * @type {Boolean}
      * @default true
      */
-    enforceCharacterLimit: PropTypes.bool
+    enforceCharacterLimit: PropTypes.bool,
+
+    /**
+     * Whether to display the character count message in red
+     *
+     * @property warnOverLimit
+     * @type {Boolean}
+     * @default false
+     */
+    warnOverLimit: PropTypes.bool
   }
 
   static defaultProps = {
     expandable: false,
-    enforceCharacterLimit: true
+    enforceCharacterLimit: true,
+    warnOverLimit: false
   }
 
   /**
@@ -154,6 +164,30 @@ class Textarea extends React.Component {
   }
 
   /**
+   * Uses the textAreaClasses method to add additional classes.
+   *
+   * @method textAreaClasses
+   * @return {String} textarea className
+   */
+  get textAreaClasses() {
+    return classNames (
+      'carbon-textarea__character-limit',
+      { 'over-limit': this.props.warnOverLimit && this.overLimit }
+    );
+  }
+
+  /**
+   * Returns if the character count exceeds the max
+   *
+   * @method overLimit
+   * @return {Boolean}
+   */
+  get overLimit() {
+    const value = this.props.value || '';
+    return value.length > parseInt(this.props.characterLimit);
+  }
+
+  /**
    * A getter that combines props passed down from the input decorator with
    * textbox specific props.
    *
@@ -200,12 +234,14 @@ class Textarea extends React.Component {
    * @return {JSX}
    */
   get characterCount() {
+    const value = this.props.value || '';
+
     if (this.props.characterLimit) {
       return (
-        <div className='carbon-textarea__character-limit' data-element='character-limit'>
+        <div className={ this.textAreaClasses } data-element='character-limit'>
           { I18n.t('textarea.limit.prefix', { defaultValue: 'You have used ' } ) }
           <span className='carbon-textarea__limit-used'>
-            { I18n.toNumber(calculateCharacterCount(this.props.value), this.i18nNumberOpts) }
+            { I18n.toNumber(value.length, this.i18nNumberOpts) }
           </span>
           { I18n.t('textarea.limit.middle', { defaultValue: ' of ' } ) }
           <span className='carbon-textarea__limit-max'>
@@ -239,13 +275,5 @@ class Textarea extends React.Component {
   }
 }
 )));
-
-let calculateCharacterCount = (value) => {
-  if (!value) { return 0; }
-
-  let limitUsed = value.length.toString(),
-      numberOfLineBreaks = (value.match(/\n/g) || []).length;
-  return parseInt(limitUsed) + numberOfLineBreaks;
-};
 
 export default Textarea;