+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# 1.1.0
+
+## Component Enhancements
+
+* `Textarea` now accepts a new prop `warnOverLimit` to display the character count message in red.
+* Simplify character count in `Textarea`.
+
+## Bug Fixes
+
+* `Date`: fixed the warning about an uncontrolled input component
+* Stop input value being removed from props (fixes Button Toggle issue)
+
 # 1.0.0
 
 ## Package Name Change
@@ -88,6 +100,7 @@ This should hopefully replace all uses of `Request` or `axios`.
 * `Link`: CSS inheritance has been updated to better support buttons.
 * `MenuList`: item filter search icon positioning is fixed
 * Row clones children when mutating props rather than creating new element to retain refs
+* Fix presence validator bug validating value as false if no props sent to validator.
 
 ## Deprecations Added